requirements.txt
astroid==2.4.2
colorama==0.4.3
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.1
importlib-metadata==2.0.0
isort==5.5.4
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
mysqlclient==2.0.1
Pillow==7.2.0
Pygments==2.7.1
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
zipp==3.3.0
